+++ src/_codemirror/themes/monokai.css.d.ts
@@ -1,3 +1,0 @@
-import {CSSResult} from 'lit-element';
-declare const codemirror: CSSResult;
-export default codemirror;